Chatgptwithtwist/calculator_rag/requirements.txt
mcp[cli]
mysql-connector-python
sentence-transformers
scikit-learn
numpy